removed central button from mobile, seperated mobile components
